(mu) use pass and independently of emacs
Instead of using standalone `gpg` keys, it is much simpler to
put all the credentials in a `pass` password store.
The commit changes the `PassCmd` commands to run when synching
the database of emails and also adapts the documentation.
This commits also removes the instadeep account as it is disabled
and useless now.