throw an error when the user choose to exit
This commit also changes the PWD of the user
when jumping to editing a config file, e.g.
`vcfg` followed by "nufu" will open the
`~/.config/nushell/scripts/functions.nu` script
but also move the user to the `~/.config/nushell/scripts/`
directory to have access to the `nushell` scripts.

When the user exits the editor, the directory is
recovered.